[FIX] base_report_to_printer: do not download text report sent to printer
In the case of a report of type "qweb-text" set to "Send to printer" (e.g. a zpl report), the pervious implementation was triggering a download of the report.

This is because inside the base odoo report handling function (odoo/addons/web/static/src/webclient/actions/action_service.js -> `_executeReportAction`), the code defaults to download for text reports that have not been dealt with.
Therefore, the current implementation ensures the "qweb-text" case is dealt as the "qweb-pdf" one ensuring we do not let this case fall into default behavior.